Do not call init_logger from inside watch.
And move init_logger to main.rs for future use.
